chore(refactor): improve tests
